revert private backpacks in coop
with no good way to indicate whose backpack is whose (because .colormap
only works on players outside the original software renderer) this led to
a lot of ignored mystery backpacks players kept obsessively touching.
going out of your way for your own ammo is also a waste of time unless
the level is extremely linear, which good quake levels never are.