Adjust position of secondary lightning beams
Due to the Z component of the left vector being discarded after the call to
normalize(), the length of the resulting vector would differ based on the
original angle of elevation. Thus when the player fired the lightning gun, the
secondary beams would get closer together as the player’s view tended towards
straight up or down.

To get the secondary beams to remain a consistent distance apart, call
normalize() after applying the vector transformation, instead of beforehand.